improvements in project module

+++ src/project/project.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { Prisma, project } from '@prisma/client';
 import { NewProject, UpdateProject } from 'src/graphql.schema';
 import { PrismaService } from '../prisma/prisma.service';
@@ -7,32 +7,85 @@ import { PrismaService } from '../prisma/prisma.service';
 export class ProjectService {
   constructor(private prisma: PrismaService) {}
 
-  private projectIncludeOptions: Prisma.projectInclude = {
+  private projectIncludeSelect: Prisma.projectSelect = {
     developers: {
-      include: {
-        dev: true
+      select: {
+        dev: true,
       }
     },
     roles: {
-      include: {
+      select: {
         role: true
       }
     }
   }
 
   async findOne(id: string): Promise<project | null> {
-    return this.prisma.project.findUnique({
+    const project = await this.prisma.project.findUnique({
       where: {
         id,
       },
-      include: this.projectIncludeOptions
+      include: this.projectIncludeSelect
     });
+
+    if (project) {
+      const dev = []
+      const roles = []
+      
+      for (const developer of project['developers'] ) {
+        dev.push(developer["dev"])
+      }
+      
+      for (const role of project['roles'] ) {
+        roles.push(role)
+      }
+      
+      const response = {
+        id: project.id,
+        name: project.name,
+        description:  project.description,
+        status: project.status,
+        developers: dev,
+        roles: roles
+      };
+      return response
+    } else {
+      throw new NotFoundException('Developer Not found')
+    }
+
+    
   }
 
   async findAll(): Promise<project[]> {
-    return this.prisma.project.findMany({
-      include: this.projectIncludeOptions      
-    });
+    try {
+      const projects = await this.prisma.project.findMany({
+        include: this.projectIncludeSelect      
+      });
+      const response = []
+      for (const item of projects) {
+        const dev = []
+        const roles = []
+        for (const developer of item['developers'] ) {
+          dev.push(developer["dev"])
+        }
+        for (const role of item['roles'] ) {
+          roles.push(role)
+        }
+        response.push({
+          id: item.id,
+          name: item.name,
+          description:  item.description,
+          developers: dev,
+          roles: roles
+        });
+      }
+      
+      return response
+    } catch (error) {
+      console.log(`[PROJ.SRV]: ${error}`);      
+      throw new InternalServerErrorException("Unknown error");
+    }
+    
   }
 
   async create(input: NewProject): Promise<project> {
@@ -48,44 +101,78 @@ export class ProjectService {
     const updateProjectData: Prisma.projectUpdateInput = {
       name: params_without_id?.name,
       description: params_without_id?.description,
-      developers: {
-        connectOrCreate: {
-          where: {
-            projectId_devId: {
-              devId: params_without_id?.developerId,
-              projectId: id
+      ...(params_without_id?.developerId && {
+        developers: {
+          connectOrCreate: {
+            where: {
+              projectId_devId: {
+                devId: params_without_id?.developerId,
+                projectId: id
+              }
+            },
+            create: {
+              devId: params_without_id.developerId
             }
-          },
-          create: {
-            devId: params_without_id.developerId
           }
         }
-      },
-      roles: {
-        connectOrCreate: {
-          where: {
-            projectId_roleId: {
-              projectId: id,
+      }),
+      ...(params_without_id?.roleId && {
+        roles: {
+          connectOrCreate: {
+            where: {
+              projectId_roleId: {
+                projectId: id,
+                roleId: params_without_id?.roleId
+              }
+            },
+            create: {
               roleId: params_without_id?.roleId
             }
-          },
-          create: {
-            roleId: params_without_id?.roleId
           }
         }
-      },
+      }),
       status: params_without_id.status
     }
 
-    return this.prisma.project.update({
+    if(!await this.findOne(id)) {
+      throw new NotFoundException("Project not found");      
+    }
+
+    const updateProject = this.prisma.project.update({
       where: {
         id,
       },
       data: updateProjectData,
+      include: this.projectIncludeSelect
     });
+
+    const proj = await updateProject
+
+    const dev = []
+    const roles = []
+    
+    for (const developer of proj['developers'] ) {
+      dev.push(developer['dev'])
+    }
+    for (const role of proj['roles'] ) {
+      roles.push(role)
+    }
+
+    const response = {
+      id: proj.id,
+      name: proj.name,
+      desription:  proj.description,
+      developers: dev,
+      roles: roles
+    }    
+    
+    return response
   }
 
   async delete(id: string): Promise<project> {
+    if(!await this.findOne(id)) {
+      throw new NotFoundException("Project not found");      
+    }
     return this.prisma.project.delete({
       where: {
         id,
